+++ examples/06_customhook/package.json
@@ -1,23 +1,21 @@
 {
-  "name": "react-tracked-example",
-  "version": "0.1.0",
+  "name": "example",
+  "version": "0.0.0",
   "private": true,
+  "type": "commonjs",
   "dependencies": {
-    "react": "18.0.0-rc.2",
-    "react-dom": "18.0.0-rc.2",
+    "@types/react": "latest",
+    "@types/react-dom": "latest",
+    "react": "latest",
+    "react-dom": "latest",
+    "react-scripts": "latest",
     "react-tracked": "latest",
-    "react-scripts": "latest"
+    "typescript": "latest"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  }
 }
