+++ examples/06_customhook/public/index.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <title>react-tracked example</title>
+    <title>example</title>
   </head>
   <body>
     <div id="app"></div>
